+++ src/model/ModelBase.ts
@@ -95,7 +95,7 @@ type BaseModelInstanceKey =
   | "wasChanged" | "getChanges" | "getOriginal" | "syncOriginal" | "discardChanges"
   | "replicate" | "makeHidden" | "makeHiddenIf" | "qualifyColumn" | "qualifyColumns"
   | "makeVisible" | "makeVisibleIf" | "append" | "setAppends" | "mergeAppends" | "hasAppended" | "withoutAppends" | "getAppends" | "save" | "update" | "updateQuietly"
-  | "updateTimestamps" | "touch" | "increment" | "decrement" | "is" | "isNot" | "isInstanceOf"
+  | "updateTimestamps" | "touch" | "increment" | "incrementQuietly" | "decrement" | "decrementQuietly" | "is" | "isNot" | "isInstanceOf"
   | "load" | "loadMissing" | "loadMorph" | "loadCount" | "loadSum" | "loadAvg" | "loadMin"
   | "loadMax" | "delete" | "saveQuietly" | "deleteQuietly" | "restore" | "restoreQuietly"
   | "forceDelete" | "forceDeleteQuietly" | "fresh" | "refresh" | "toJSON" | "json" | "toString" | "freshTimestamp"